[Bronze III] Title: 주차 요금 정산하기, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/33753. 주차 요금 정산하기/주차 요금 정산하기.cc
@@ -9,8 +9,7 @@ int main() {
     t-=30;
     
     if(t >= 0) {
-        s = c * t/b;
-        
+        s = c * (int)(t/b);
         if(t%b != 0) s += c;
     }
     cout << a + s;
